Second iteration of light palette
This commit tackles:

- Lualine colors.
Lualine had been made less contrasting in order
to fit both light and dark backgrounds.
Specifically, use of colors above nord7
was restricted to cases where text is not
as important as the visual indication of the
mode we're in.

- TabLine colors
TabLine colors are now in agreement with lualine
colors and should look fitting out of the box

- Visual selection
Visual selection is now a shade more intence than
cursor line, making it noticable both in dark
and light mode.

- Lualine inactive colors
A typo that snuck in some time ago was fixed too,
lualine inactive colors are back.