Handling error
I had this problem and it took me a while to resolve it. When the 'hi clear' command is used in a nord theme function, it makes the Lualine's background disappear. After I found that, I made the lualine file be called only after nord.nvim had already run. For this reason, I think this warning is important, because there may be several people who will experience the same error.